project/requirements.txt
Flask
Flask-SQLAlchemy
Flask-WTF
python-dotenv
pymysql
mysqlclient